Rename run_tests.sh build_tests.sh
I think I want the test system to be as modular as possible, so there should be
separate scripts for building, running, and comparing the output (plus a PBS
script so the running / output comparison occur on compute nodes). This will
require a little bit of a refactor, and I'm tempted to do this all in Python.